+++ skins/desktop/skin.php
@@ -101,7 +101,7 @@
 				<div id="lastfm" class="invisible topstats">
 					<?php
 					print '<i title="'.language::gettext('button_love').
-						'" class="icon-heart npicon clickicon tooltip" id="love"></i>';
+						'" class="icon-heart npicon clickicon tooltip spinable" id="love"></i>';
 					?>
 				</div>
 				<div id="playcount" class="topstats"></div>