+++ skins/fruit/uibits.class.php
@@ -1,3 +1,0 @@
-<?php
-include ("skins/standard_ui.php");
-?>