+++ skins/standard_ui.php
@@ -73,7 +73,7 @@ class uibits {
 
 		// Menu Button
 		if ($data['ttid']) {
-			$button_class = "icon-menu inline-icon fixed clickable clickicon invisibleicon clicktrackmenu";
+			$button_class = "icon-menu inline-icon fixed clickable clickicon invisibleicon clicktrackmenu spinable";
 			if ($data['lm'] === null) {
 				$button_class .= ' clickremovedb';
 			}