+++ skins/tablet/skinvars.php
@@ -1,6 +1,0 @@
-<?php
-$use_plugins = true;
-$only_plugins_on_menu = false;
-$use_smartradio = true;
-$snapcast_in_volume = true;
-?>